web/requirements.txt
#install using  'pip install -r requirements.txt'

Django >= 1.10.2, < 2
uWSGI==2.0.14

#django-two-factor-auth==1.4.0
django_otp==0.3.7
django-phonenumber-field==0.7.2
qrcode==4.0.4
django-formtools==1.0

#django-user-sessions==1.3.1
django-redis-sessions==0.5.0

greenlet==0.4.10
gevent==1.1.2
w1thermsensor==0.3.2
django-websocket-redis==0.4.7
pytz==2016.7
RPi.GPIO==0.6.3
